hg commit -m "I think it's feature complete and optimized. Added pwm bitmaps, pwm takes about 1.5us, commutations about 2.5, motor about 1.5. It runs ok jacked up to 250khz, but for my application there's no real need and I prefer the interrupts to be more regular, especially since I'm planning on trying to initially run without feedback (ha!).

+++ Motor.h
@@ -14,8 +14,8 @@ class Motor {
     int poles; 
     
   private:
-    long _ticks = 0;
-    int _ticks_per_phase;
+    unsigned int _ticks = 0;
+    unsigned int _ticks_per_phase;
     
 };
